google_adk/requirements.txt
google-adk
langchain
langchain-community
sentence-transformers
pymilvus
uvicorn
fastapi
sqlalchemy
psycopg2-binary
python-dotenv
PyJWT
